Completed Math utilities

+++ src/main/java/com/zipcodewilmington/danny_do_better_exercises/MathUtilities.java
@@ -29,7 +29,7 @@ public class MathUtilities {
      * @return sum of `baseValue` and `difference`
      */
     public Short add(short baseValue, short difference) {
-        return baseValue + difference;
+        return (short) (baseValue + difference);
     }
 
     /**
@@ -38,7 +38,7 @@ public class MathUtilities {
      * @return sum of `baseValue` and `difference`
      */
     public Byte add(byte baseValue, byte difference) {
-        return baseValue + difference;
+        return (byte)(baseValue + difference);
     }
 
     /**
@@ -83,7 +83,7 @@ public class MathUtilities {
      * @return difference between `baseValue` and `difference`
      */
     public Short subtract(short baseValue, short difference) {
-        return baseValue - difference;
+        return (short)(baseValue - difference);
     }
 
     /**
@@ -92,7 +92,7 @@ public class MathUtilities {
      * @return difference between `baseValue` and `difference`
      */
     public Byte subtract(byte baseValue, byte difference) {
-        return baseValue - difference;
+        return (byte)(baseValue - difference);
     }
 
     /**
@@ -138,7 +138,7 @@ public class MathUtilities {
      * @return division of `dividend` by `divisor
      */
     public Short divide(short dividend, short divisor) {
-        return dividend / divisor;
+        return (short)(dividend / divisor);
     }
 
     /**
@@ -147,7 +147,7 @@ public class MathUtilities {
      * @return division of `dividend` by `divisor
      */
     public Byte divide(byte dividend, byte divisor) {
-        return dividend / divisor;
+        return (byte)(dividend / divisor);
     }
 
     /**
@@ -193,7 +193,8 @@ public class MathUtilities {
      * @return product of `multiplicand` by `multiplier`
      */
     public Short multiply(short multiplicand, short multiplier) {
-        return multiplicand * multiplier;
+
+        return (short)(multiplicand * multiplier);
     }
     /**
      * @param multiplicand value to be multiplied
@@ -201,7 +202,7 @@ public class MathUtilities {
      * @return product of `multiplicand` by `multiplier`
      */
     public Byte multiply(byte multiplicand, byte multiplier) {
-        return multiplicand * multiplier;
+        return (byte)(multiplicand * multiplier);
     }
 
     /**
